MySubjects: небольшие улучшения кода
