+++ src/features/teacher-subjects/components/teacher-subjects-toolbar.tsx
@@ -11,8 +11,7 @@ interface TeacherSubjectsToolbarProps {
   filterType: SubjectFilterType;
   onFilterTypeChange: (value: SubjectFilterType) => void;
   onAddSubject: () => void;
-  isAddingRow: boolean;
-  rowsExist: boolean;
+  showAddButton: boolean;
 }
 
 export function TeacherSubjectsToolbar({
@@ -21,8 +20,7 @@ export function TeacherSubjectsToolbar({
   filterType,
   onFilterTypeChange,
   onAddSubject,
-  isAddingRow,
-  rowsExist,
+  showAddButton,
 }: TeacherSubjectsToolbarProps) {
   return (
     <div className="flex flex-wrap items-center gap-3">
@@ -43,10 +41,10 @@ export function TeacherSubjectsToolbar({
         size="sm"
       />
 
-      {!isAddingRow && rowsExist ? (
+      {showAddButton ? (
         <Button className="ml-auto" onClick={onAddSubject}>
           <Plus className="size-4" data-icon="inline-start" />
-          {"Добавить предмет"}
+            Добавить предмет
         </Button>
       ) : null}
     </div>
