+++ src/lib/types.ts
@@ -1,5 +1,4 @@
 import type { SubjectType } from "@/generated/prisma/client";
-import { ReactNode } from "react";
 
 export type SubjectRenderInfo = {
   value: SubjectType;
@@ -10,11 +9,3 @@ export type SubjectRenderInfo = {
 };
 
 export type SubjectFilterType = "ALL" | SubjectType;
-
-export type EmptyStateConfig = {
-  title: string;
-  description?: ReactNode;
-  icon?: ReactNode;
-  action?: ReactNode;
-  className?: string;
-};
